% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galvanizer-package.R
\docType{package}
\name{galvanizer-package}
\alias{galvanizer}
\alias{galvanizer-package}
\title{galvanizer: Interface to Galvanize 'Highbond' Internal Audit Software}
\description{
An R interface to the Galvanize 'Highbond' API <https://docs-apis.highbond.com>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonlinca.github.io/galvanizer/}
  \item \url{https://github.com/jonlinca/galvanizer}
  \item Report bugs at \url{https://github.com/jonlinca/galvanizer/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Lin \email{hello@jonlin.ca} (\href{https://orcid.org/0000-0001-8664-6846}{ORCID})

}
\keyword{internal}
