% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_results.R
\name{get_highbond_results}
\alias{get_highbond_results}
\title{Retrieve Highbond Results table}
\usage{
get_highbond_results(
  apikey,
  org,
  datacenter,
  table_id,
  timezone = Sys.timezone()
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{One of \code{'us', 'ca', 'eu', 'ap', 'au'}}

\item{table_id}{The table to be downloaded}

\item{timezone}{Defaults to \code{Sys.timezone()}. See \code{OlsonNames()} for a list of allowable timezones.}
}
\value{
A list containing the contents, table requested and the original response
}
\description{
Downloads the content in the structure as saved by Highbond
Results, including metadata and questionnaire responses (if applicable).
}
\details{
Table ID can be found in the path of your Result table.
https://<company_name>.results.highbond.com/projects/<COLLECTION_ID>/controls/<ANALYSIS_ID>/control_tests/<TABLE_ID>

Requires a Highbond API token. Activate a token from
https://help.highbond.com/helpdocs/highbond/en-us/Content/launchpad/getting_started/managing_access_tokens.html.

The Organization number and datacenter can both be found from the Highbond launchpad, then accessing Options and Organization.
https://accounts.highbond.com/orgs/<ORG_ID>/details
}
\examples{
\dontrun{
response <- get_highbond_results(highbond_openapi, 12345, 'us', 567890)
}
}
