% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_control_test_plan}
\alias{get_project_control_test_plan}
\title{Retrieve Highbond Project - Test Plans}
\usage{
get_project_control_test_plan(
  apikey,
  org,
  datacenter,
  control_test_plan_id,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{control_test_plan_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of control test plans
}
\description{
"A test plan is a document that details how controls are
assessed. Test plans identify the testing method or type of evidence
obtained, specify the total sample size (split amongst testing rounds), and
illustrate test steps or attributes."
}
\details{
Only applicable to test plans with Internal Control types

possible fields: testing_round_number, not_applicable, sample_size,
testing_results, testing_conclusion, created_at, updated_at, control,
assigned_user
}
