% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_control_tests}
\alias{get_project_control_tests}
\title{Retrieve Highbond Project - Control Tests}
\usage{
get_project_control_tests(
  apikey,
  org,
  datacenter,
  control_test_id,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{control_test_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of control tests
}
\description{
"Control tests evaluate the operating effectiveness of a control."
}
\details{
Only applicable to test plans with Internal Control types
}
