% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_entities}
\alias{get_project_entities}
\title{Retrieve Highbond Projects - Entities}
\usage{
get_project_entities(
  auth,
  entity_id = NULL,
  fields = NULL,
  pagesize = 50,
  waittime = 0.2
)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{entity_id}{\code{NULL} will default to all entities.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}

\item{waittime}{Time in seconds to wait between requests.}
}
\value{
A tibble of entities
}
\description{
Get all the entities
}
\details{
Fields allowed: title, description, created_at, updated_at, parent, children_count, entity_category
}
