% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_entity_categories}
\alias{get_project_entity_categories}
\title{Retrieve Highbond Projects - Entity Categories}
\usage{
get_project_entity_categories(
  auth,
  entity_category_id = NULL,
  fields = NULL,
  pagesize = 50,
  waittime = 0.2
)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{entity_category_id}{\code{NULL} will default to all entity categories}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}

\item{waittime}{Time in seconds to wait between requests.}
}
\value{
A tibble of entities
}
\description{
Get all the entity categories
}
\details{
Fields allowed: title, description, position, entities_count, created_at, updated_at, entities
}
