% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_mitigation}
\alias{get_project_mitigation}
\title{Retrieve Highbond Project - Mitigations}
\usage{
get_project_mitigation(
  apikey,
  org,
  datacenter,
  mitigation_id,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{mitigation_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of mitigations
}
\description{
The mapping between controls and risks
}
