% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_mitigations}
\alias{get_project_mitigations}
\title{Retrieve Highbond Project - Mitigations}
\usage{
get_project_mitigations(
  auth,
  mitigation_id,
  fields = NULL,
  pagesize = 50,
  waittime = 0.2
)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{mitigation_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}

\item{waittime}{Time in seconds to wait between requests.}
}
\value{
A tibble of mitigations
}
\description{
The mapping between controls and risks
}
