% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_planning_file}
\alias{get_project_planning_file}
\title{Retrieve Highbond Project - Planning Files}
\usage{
get_project_planning_file(
  apikey,
  org,
  datacenter,
  project_id,
  planning_file_id = NULL,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{project_id}{Project ID number. Required.}

\item{planning_file_id}{Defaults to all planning files within a project. NULL
defaults to all planning files}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of planning files.
}
\description{
Downloads the primary details of one or multiple planning files
for a project.
}
\details{
possible fields: name, reference_id, description, position,
created_at, updated_at, custom_attributes, project
}
