% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_risks}
\alias{get_project_risks}
\title{Retrieve Highbond Project - Risks}
\usage{
get_project_risks(
  auth,
  objective_id = NULL,
  risk_id = NULL,
  fields = NULL,
  pagesize = 50,
  waittime = 0.2
)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{objective_id}{Required if other parameter is blank. May obtain multiple rows.}

\item{risk_id}{Required if other parameter is blank. Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}

\item{waittime}{Time in seconds to wait between requests.}
}
\value{
A tibble of Risks
}
\description{
"A narrative is a description of a business process or area
under review. Narratives are also known as policies, process narratives,
process descriptions, or control guides."
}
\details{
possible fields: title, description, risk_id, position, impact,
likelihood, custom_attributes, created_at, updated_at, objective,
mitigations, entities
}
