% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_type}
\alias{get_project_type}
\title{Retrieve Highbond Projects - Project Types}
\usage{
get_project_type(
  apikey,
  org,
  datacenter,
  project_type_id = NULL,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{project_type_id}{Defaults to all project types}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of project types
}
\description{
Downloads the primary details of one or all projects
}
\details{
Possible fields include:
}
