% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_walkthrough}
\alias{get_project_walkthrough}
\title{Retrieve Highbond Project - Walkthroughs / Execute Procedures}
\usage{
get_project_walkthrough(
  apikey,
  org,
  datacenter,
  walkthrough_id,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{walkthrough_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of walkthroughs
}
\description{
"A walkthrough is a series of steps you perform to establish the
reliability of controls and test the design of controls. Each control you
define has a corresponding walkthrough that is used to verify that the
control is designed appropriately. In a Workplan workflow project, a
walkthrough is called an execute procedure."
}
