% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results_analyses}
\alias{get_results_analyses}
\title{Retrieve Highbond Results - Analyses}
\usage{
get_results_analyses(auth, collection_id = NULL, analysis_id = NULL)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{collection_id}{Collection ID. Required if other parameter is blank.}

\item{analysis_id}{Analyses ID. Required if other parameter is blank.}
}
\value{
A data frame of Analyses in a Collection
}
\description{
Downloads a list of Analyses in a Collection
}
