% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results_collection}
\alias{get_results_collection}
\title{Retrieve Highbond Results - Collections}
\usage{
get_results_collection(auth, collection_id = NULL)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{collection_id}{Collection number. \code{NULL} will default to all collections.}
}
\value{
A tibble of collections
}
\description{
Downloads the primary details of one or all collections
}
