% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results_collections}
\alias{get_results_collections}
\title{Retrieve Highbond Results - Collections}
\usage{
get_results_collections(auth, collection_id = NULL)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{collection_id}{Collection ID. Optional.}
}
\value{
A data frame of Collections
}
\description{
Downloads a list of collections
}
