% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_results_tables}
\alias{get_results_tables}
\title{Retrieve Highbond Results - Tables}
\usage{
get_results_tables(auth, analysis_id = NULL, table_id = NULL)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{analysis_id}{Analysis ID. Required if other parameter is blank.}

\item{table_id}{Table ID. Required if other parameter is blank.}
}
\value{
A data frame of Tables in an Analyses
}
\description{
Downloads the primary details of one or all tables in an Analyses
}
