% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_control_tests}
\alias{get_project_control_tests}
\title{Retrieve Highbond Project - Control Tests}
\usage{
get_project_control_tests(auth, control_test_id, fields = NULL, pagesize = 50)
}
\arguments{
\item{auth}{Highbond authentication credentials, created from \code{\link{setup_highbond}}}

\item{control_test_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of control tests
}
\description{
"Control tests evaluate the operating effectiveness of a control."
}
\details{
Only applicable to test plans with Internal Control types
}
