% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_narrative}
\alias{get_project_narrative}
\title{Retrieve Highbond Project - Narratives}
\usage{
get_project_narrative(
  apikey,
  org,
  datacenter,
  objective_id = NULL,
  narrative_id = NULL,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{objective_id}{Required if other parameter is blank. May obtain multiple rows.}

\item{narrative_id}{Required if other parameter is blank. Will get only one row.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of narratives
}
\description{
"A narrative is a description of a business process or area under
review. Narratives are also known as policies, process narratives, process
descriptions, or control guides."
}
\details{
possible fields: title, description, created_at, updated_at, objective
}
