% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_planning_file}
\alias{get_project_planning_file}
\title{Retrieve Highbond Project Planning Files}
\usage{
get_project_planning_file(
  apikey,
  org,
  datacenter,
  project_id,
  planning_file_id = NULL,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{Can be us, ca, eu, ap, au}

\item{project_id}{Defaults to all projects}

\item{planning_file_id}{Defaults to all planning files within a project}

\item{fields}{OPTIONAL. A character string of all fields requested within the
project}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A dataframe of project types
}
\description{
Downloads the primary details of one or multiple planning files for a project.
}
\details{
possible fields:
}
