% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_request_items}
\alias{get_project_request_items}
\title{Retrieve Highbond Project - Request Items}
\usage{
get_project_request_items(
  apikey,
  org,
  datacenter,
  request_id,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{One of \code{'us', 'ca', 'eu', 'ap', 'au'}}

\item{request_id}{Will get only one.}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of walkthroughs
}
\description{
Capture what you asked for
}
