% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highbond_api_projects.R
\name{get_project_type}
\alias{get_project_type}
\title{Retrieve Highbond Projects - Project Types}
\usage{
get_project_type(
  apikey,
  org,
  datacenter,
  project_type_id = NULL,
  fields = NULL,
  pagesize = 50
)
}
\arguments{
\item{apikey}{Highbond API token}

\item{org}{Organization number}

\item{datacenter}{One of \code{'us', 'ca', 'eu', 'ap', 'au'}}

\item{project_type_id}{\code{NULL} Defaults to all project types}

\item{fields}{OPTIONAL. A character vector each field requested within the
project. NULL will default to all fields.}

\item{pagesize}{Defaults to 50. Maximum is 100.}
}
\value{
A tibble of project types
}
\description{
Downloads the primary details of one or all project types.
}
\details{
Possible fields include: name, description, workflow, project_terms,
certification_terms, control_terms, finding_terms, control_test_terms,
finding_action_terms, narrative_terms, objective_terms, planning_terms,
results_terms, risk_terms, test_plan_terms, walkthrough_terms
}
